# swagger-dev.json
{
  "swagger": "2.0",
  "info": {
    "title": "CSE 341 Ecommerce",
    "description": "Final Project",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "tags": [
    {
      "name": "Login",
      "description": "Operations related to user login"
    },
    {
      "name": "Users",
      "description": "Operations related to users"
    },
    {
      "name": "Stores",
      "description": "Operations related to stores"
    },
    {
      "name": "Products",
      "description": "Operations related to products"
    },
    {
      "name": "Cart",
      "description": "Operations related to shopping cart"
    },
    {
      "name": "Orders",
      "description": "Operations related to orders"
    }
    
  ],
  "paths": {
    "/login": {
      "get": {
        "tags": [
          "Login"
        ],
        "description": "User login",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "header",
            "type": "string",
            "description": "User's email address",
            "example": "user@example.com"
          },
          {
            "name": "password",
            "in": "header",
            "type": "string",
            "description": "User's password",
            "example": "password123"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "token": "jwt-token"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/stores/": {
      "get": {
        "tags": [
          "Stores"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Stores"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Store Name"
                },
                "address": {
                  "type": "string",
                  "example": "123 Main St"
                },
                "phone": {
                  "type": "string",
                  "example": "123-456-7890"
                },
                "email": {
                  "type": "string",
                  "example": "store@example.com"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "open",
                    "closed"
                  ],
                  "example": "open"
                }
              },
              "required": [
                "name",
                "address",
                "phone",
                "email",
                "status"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/stores/{id}": {
      "get": {
        "tags": [
          "Stores"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Stores"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Store Name"
                },
                "address": {
                  "type": "string",
                  "example": "123 Main St"
                },
                "phone": {
                  "type": "string",
                  "example": "123-456-7890"
                },
                "email": {
                  "type": "string",
                  "example": "store@example.com"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "open",
                    "closed"
                  ],
                  "example": "open"
                }
              },
              "required": [
                "name",
                "address",
                "phone",
                "email",
                "status"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Stores"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer token for authentication"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "John"
                },
                "last_name": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "john.doe@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "123password"
                },
                "store_id": {
                  "type": "string"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "admin",
                    "customer",
                    "manager"
                  ],
                  "example": "customer"
                }
              },
              "required": [
                "name",
                "last_name",
                "email",
                "role",
                "store_id",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer token for authentication"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer token for authentication"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "John"
                },
                "last_name": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "john.doe@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "123password"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "admin",
                    "customer",
                    "manager"
                  ],
                  "example": "customer"
                }
              },
              "required": [
                "name",
                "last_name",
                "email",
                "role",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer token for authentication"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/products/": {
      "tags": [
        "Products"
      ],
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Retrieve all products.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": [
                {
                  "_id": "60c72b2f4f1a4a3d2e75b9d3",
                  "name": "Product Name",
                  "stock": 100,
                  "description": "Product Description",
                  "brand": "Brand Name",
                  "category": "Category Name"
                }
              ]
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Create a new product.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Product Name"
                },
                "stock": {
                  "type": "integer",
                  "example": 100
                },
                "description": {
                  "type": "string",
                  "example": "Product Description"
                },
                "brand": {
                  "type": "string",
                  "example": "Brand Name"
                },
                "category": {
                  "type": "string",
                  "example": "Category Name"
                }
              },
              "required": [
                "name",
                "stock",
                "description",
                "brand",
                "category"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "examples": {
              "application/json": {
                "acknowledged": true,
                "insertedId": "60c72b2f4f1a4a3d2e75b9d3"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Retrieve a single product by its ID.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "_id": "60c72b2f4f1a4a3d2e75b9d3",
                "name": "Product Name",
                "stock": 100,
                "description": "Product Description",
                "brand": "Brand Name",
                "category": "Category Name"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Delete a product by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Update a product by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Updated Product Name"
                },
                "stock": {
                  "type": "integer",
                  "example": 150
                },
                "description": {
                  "type": "string",
                  "example": "Updated Product Description"
                },
                "brand": {
                  "type": "string",
                  "example": "Updated Brand Name"
                },
                "category": {
                  "type": "string",
                  "example": "Updated Category Name"
                }
              },
              "required": [
                "name",
                "stock",
                "description",
                "brand",
                "category"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Product Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/get/low-stock": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Retrieve all products with stock lower than 20.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": [
                {
                  "_id": "60c72b2f4f1a4a3d2e75b9d3",
                  "name": "Low Stock Product",
                  "stock": 10,
                  "description": "Product with low stock",
                  "brand": "Brand Name",
                  "category": "Category Name"
                }
              ]
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/cart/": {
      "get": {
        "tags": [
          "Cart"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Cart"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "string",
                  "example": "User ID"
                },
                "items": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "product_id": {
                        "type": "string",
                        "example": "Product ID"
                      },
                      "quantity": {
                        "type": "integer",
                        "example": 1
                      }
                    }
                  }
                }
              },
              "required": [
                "user_id",
                "items"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/cart/{id}": {
      "get": {
        "tags": [
          "Cart"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Cart"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Cart"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/cart/{id}/total": {
      "get": {
        "tags": [
          "Cart"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "string",
                  "example": "any"
                },
                "amount": {
                  "type": "number",
                  "example": 100.0
                },
                "status": {
                  "type": "string",
                  "example": "any"
                },
                "date": {
                  "type": "string",
                  "example": "any"
                },
                "cart_id": {
                  "type": "string",
                  "example": "any"
                }
              },
              "required": [
                "amount",
                "cart_id"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "string",
                  "example": "any"
                },
                "amount": {
                  "type": "number",
                  "example": 100.0
                },
                "status": {
                  "type": "string",
                  "example": "any"
                },
                "date": {
                  "type": "string",
                  "example": "any"
                },
                "cart_id": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}